+++ crates/api/src/properties/widget/mod.rs
@@ -1,10 +1,12 @@
 // Widget related properties.
+pub use self::focus_state::*;
 pub use self::keyboard_state::*;
 pub use self::render_pipeline::*;
 pub use self::selected_entities::*;
 pub use self::selected_indices::*;
 pub use self::text_selection::*;
 
+mod focus_state;
 mod keyboard_state;
 mod render_pipeline;
 mod selected_entities;